Modify LightCurve class to allow creation of feature extractor classes
